feat|fix: threadをソートとページングの機能追加

+++ app/models/thread.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
 
+from app.config import current_config
 from app.models import Base, row_to_dict, session_scope
 from app.models.category import Category
 
@@ -37,6 +38,9 @@ class Thread(Base):
                 cls.thread_id == thread_id
             ).first()
 
+            if not t_rows:
+                return None
+
             # thread_idに紐づくcommentリスト取得
             comments = Comment.get_all_by_t_id(thread_id)
 
@@ -59,17 +63,48 @@ class Thread(Base):
             return result
 
     @classmethod
-    def get_all_by_c_id(cls, category_id):
+    def get_all_by_c_id(cls, category_id, sort_id, paging):
         '''category_idに紐づくthreadリスト取得
         '''
         with session_scope() as session:
-            rows = session.query(
+            query = session.query(
                 cls
             ).join(
                 Category, Thread.category_id == Category.category_id
             ).filter(
                 Category.category_id == category_id
-            ).all()
+            )
+
+            # sort
+            if sort_id == current_config().get('ID_ASC'):
+                # ID昇順
+                pass
+            elif sort_id == current_config().get('ID_DESC'):
+                # ID降順
+                query = query.order_by(cls.thread_id.desc())
+            elif sort_id == current_config().get('SPEED_DESC'):
+                # 人気高い順
+                query = query.order_by(cls.speed.desc())
+            elif sort_id == current_config().get('SPEED_ASC'):
+                # 人気低い順
+                query = query.order_by(cls.speed.asc())
+            elif sort_id == current_config().get('NUM_COMMENT_DESC'):
+                # コメント数多い順
+                query = query.order_by(cls.thread_id.desc())
+            elif sort_id == current_config().get('NUM_COMMENT_ASC'):
+                # コメント数少ない順
+                query = query.order_by(cls.thread_id.asc())
+
+            # paging
+            # ex)
+            # 1. offset=0, limit=10
+            # 2. offset=10, limit=20
+            offset = (paging - 1) * 10
+            limit = 10
+
+            query = query.offset(offset).limit(limit)
+
+            rows = query.all()
 
             result = [row_to_dict(row) for row in rows]
 
@@ -85,12 +120,25 @@ class Thread(Base):
             return row_to_dict(data)
 
     @classmethod
-    def delete(cls, params):
-        with session_scope() as session:
-            thread_id = params['thread_id']
+    def delete(cls, thread_id):
+        from app.models.comment import Comment
 
+        with session_scope() as session:
+            # thread_idに紐づくcomment削除
+            rows = session.query(
+                Comment
+            ).filter(
+                cls.thread_id == thread_id
+            ).all()
+
+            if not rows:
+                return
+
+            # thread_idに紐づくcommentの削除
+            for row in rows:
+                session.delete(row)
+
+            # threadの削除
             row = session.query(cls).filter_by(thread_id=thread_id).first()
 
             session.delete(row)
-
-            # *** TODO thread_idに紐づくcomment削除
